Command line to generate a FLV from TS file.
Modify to allow running the test code (shamelessly lifted from the
muxer test HTML) from the command line (nodeJS) to generate a FLV from
a source TS file.

If people like this, we can clean it up and add it to the Gruntfile.js.
 Note some talk about how to make the src files more require/export
friendly (a moving target I know) would be in order first.

To run it now use:
   node ./test/muxer/js/tsToFLV.js ~/mydir/myfile.ts  outfile.flv

It will generate outfile.flv in ~/mydir